split the bulk of SerialExecutor into BaseExecutor for reuse
don't test the abstract class separately though. When we add the
second implementation, we'll split the bulk of the specs into a shared
context
